Remove type id based implicit casting
