+++ src/binder/bind_expression/bind_boolean_expression.cpp
@@ -21,7 +21,7 @@ std::shared_ptr<Expression> ExpressionBinder::bindBooleanExpression(ExpressionTy
     const expression_vector& children) {
     expression_vector childrenAfterCast;
     for (auto& child : children) {
-        childrenAfterCast.push_back(implicitCastIfNecessary(child, LogicalTypeID::BOOL));
+        childrenAfterCast.push_back(implicitCastIfNecessary(child, *LogicalType::BOOL()));
     }
     auto functionName = expressionTypeToString(expressionType);
     function::scalar_func_exec_t execFunc;
