+++ src/binder/bind_expression/bind_case_expression.cpp
@@ -58,7 +58,7 @@ std::shared_ptr<Expression> ExpressionBinder::bindCaseExpression(
         for (auto i = 0u; i < parsedCaseExpression.getNumCaseAlternative(); ++i) {
             auto caseAlternative = parsedCaseExpression.getCaseAlternative(i);
             auto boundWhen = bindExpression(*caseAlternative->whenExpression);
-            boundWhen = implicitCastIfNecessary(boundWhen, LogicalTypeID::BOOL);
+            boundWhen = implicitCastIfNecessary(boundWhen, *LogicalType::BOOL());
             auto boundThen = bindExpression(*caseAlternative->thenExpression);
             boundThen = implicitCastIfNecessary(boundThen, resultType);
             boundCaseExpression->addCaseAlternative(boundWhen, boundThen);
