+++ src/function/list/CMakeLists.txt
@@ -1,6 +1,19 @@
 add_library(kuzu_list_function
         OBJECT
-        list_to_string_function.cpp)
+        list_agg_function.cpp
+        list_any_value_function.cpp
+        list_binary_right_switch_function.cpp
+        list_concat_function.cpp
+        list_creation.cpp
+        list_distinct_function.cpp
+        list_extract_function.cpp
+        list_range_function.cpp
+        list_reverse_function.cpp
+        list_slice_function.cpp
+        list_sort_function.cpp
+        list_to_string_function.cpp
+        list_unique_function.cpp
+        size_function.cpp)
 
 set(ALL_OBJECT_FILES
         ${ALL_OBJECT_FILES} $<TARGET_OBJECTS:kuzu_list_function>
