+++ src/function/list/list_creation.cpp
@@ -1,0 +1,57 @@
+#include "binder/expression/expression_util.h"
+#include "function/list/vector_list_functions.h"
+#include "function/scalar_function.h"
+
+using namespace kuzu::common;
+
+namespace kuzu {
+namespace function {
+
+void ListCreationFunction::execFunc(const std::vector<std::shared_ptr<ValueVector>>& parameters,
+    ValueVector& result, void* /*dataPtr*/) {
+    result.resetAuxiliaryBuffer();
+    for (auto selectedPos = 0u; selectedPos < result.state->selVector->selectedSize;
+         ++selectedPos) {
+        auto pos = result.state->selVector->selectedPositions[selectedPos];
+        auto resultEntry = ListVector::addList(&result, parameters.size());
+        result.setValue(pos, resultEntry);
+        auto resultDataVector = ListVector::getDataVector(&result);
+        auto resultPos = resultEntry.offset;
+        for (auto i = 0u; i < parameters.size(); i++) {
+            const auto& parameter = parameters[i];
+            auto paramPos = parameter->state->isFlat() ?
+                                parameter->state->selVector->selectedPositions[0] :
+                                pos;
+            resultDataVector->copyFromVectorData(resultPos++, parameter.get(), paramPos);
+        }
+    }
+}
+
+static std::unique_ptr<FunctionBindData> bindFunc(const binder::expression_vector& arguments,
+    Function* /*function*/) {
+    LogicalType combinedType(LogicalTypeID::ANY);
+    binder::ExpressionUtil::tryCombineDataType(arguments, combinedType);
+    if (combinedType.getLogicalTypeID() == LogicalTypeID::ANY) {
+        combinedType = *LogicalType::STRING();
+    }
+    auto resultType = LogicalType::LIST(combinedType.copy());
+    auto bindData = std::make_unique<FunctionBindData>(std::move(resultType));
+    for (auto& _ : arguments) {
+        (void)_;
+        bindData->paramTypes.push_back(combinedType);
+    }
+    return bindData;
+}
+
+function_set ListCreationFunction::getFunctionSet() {
+    function_set result;
+    auto function =
+        std::make_unique<ScalarFunction>(name, std::vector<LogicalTypeID>{LogicalTypeID::ANY},
+            LogicalTypeID::LIST, execFunc, nullptr, bindFunc);
+    function->isVarLength = true;
+    result.push_back(std::move(function));
+    return result;
+}
+
+} // namespace function
+} // namespace kuzu
