+++ src/function/path/CMakeLists.txt
@@ -1,0 +1,10 @@
+add_library(kuzu_function_path
+        OBJECT
+        nodes_function.cpp
+        properties_function.cpp
+        rels_function.cpp
+        semantic_function.cpp)
+
+set(ALL_OBJECT_FILES
+        ${ALL_OBJECT_FILES} $<TARGET_OBJECTS:kuzu_function_path>
+        PARENT_SCOPE)
