+++ src/function/struct/CMakeLists.txt
@@ -1,0 +1,8 @@
+add_library(kuzu_function_struct
+        OBJECT
+        struct_extract_function.cpp
+        struct_pack_function.cpp)
+
+set(ALL_OBJECT_FILES
+        ${ALL_OBJECT_FILES} $<TARGET_OBJECTS:kuzu_function_struct>
+        PARENT_SCOPE)
