+++ src/function/struct/struct_pack_function.cpp
@@ -1,18 +1,13 @@
-#include "function/struct/vector_struct_functions.h"
-
-#include "binder/expression/literal_expression.h"
-#include "binder/expression_binder.h"
-#include "common/exception/binder.h"
-#include "common/string_format.h"
 #include "function/scalar_function.h"
+#include "function/struct/vector_struct_functions.h"
 
 using namespace kuzu::common;
 
 namespace kuzu {
 namespace function {
 
-static std::unique_ptr<FunctionBindData> StructPackBindFunc(
-    const binder::expression_vector& arguments, Function* /*function*/) {
+static std::unique_ptr<FunctionBindData> bindFunc(const binder::expression_vector& arguments,
+    Function* /*function*/) {
     std::vector<StructField> fields;
     for (auto& argument : arguments) {
         if (argument->getDataType().getLogicalTypeID() == LogicalTypeID::ANY) {
@@ -21,7 +16,7 @@ static std::unique_ptr<FunctionBindData> StructPackBindFunc(
         fields.emplace_back(argument->getAlias(), argument->getDataType().copy());
     }
     auto resultType = LogicalType::STRUCT(std::move(fields));
-    return std::make_unique<FunctionBindData>(std::move(resultType));
+    return FunctionBindData::getSimpleBindData(arguments, *resultType);
 }
 
 void StructPackFunctions::compileFunc(FunctionBindData* /*bindData*/,
@@ -74,49 +69,9 @@ function_set StructPackFunctions::getFunctionSet() {
     function_set functions;
     auto function =
         std::make_unique<ScalarFunction>(name, std::vector<LogicalTypeID>{LogicalTypeID::ANY},
-            LogicalTypeID::STRUCT, execFunc, nullptr, compileFunc, StructPackBindFunc);
+            LogicalTypeID::STRUCT, execFunc, nullptr, compileFunc, bindFunc);
     function->isVarLength = true;
     functions.push_back(std::move(function));
-    return functions;
-}
-
-std::unique_ptr<FunctionBindData> StructExtractFunctions::bindFunc(
-    const binder::expression_vector& arguments, Function* /*function*/) {
-    auto structType = arguments[0]->getDataType();
-    if (arguments[1]->expressionType != ExpressionType::LITERAL) {
-        throw BinderException("Key name for struct/union extract must be STRING literal.");
-    }
-    auto key = ((binder::LiteralExpression&)*arguments[1]).getValue()->getValue<std::string>();
-    auto fieldIdx = StructType::getFieldIdx(&structType, key);
-    if (fieldIdx == INVALID_STRUCT_FIELD_IDX) {
-        throw BinderException(stringFormat("Invalid struct field name: {}.", key));
-    }
-    return std::make_unique<StructExtractBindData>(
-        (StructType::getFieldTypes(&structType))[fieldIdx]->copy(), fieldIdx);
-}
-
-void StructExtractFunctions::compileFunc(FunctionBindData* bindData,
-    const std::vector<std::shared_ptr<ValueVector>>& parameters,
-    std::shared_ptr<ValueVector>& result) {
-    KU_ASSERT(parameters[0]->dataType.getPhysicalType() == PhysicalTypeID::STRUCT);
-    auto structBindData = ku_dynamic_cast<FunctionBindData*, StructExtractBindData*>(bindData);
-    result = StructVector::getFieldVector(parameters[0].get(), structBindData->childIdx);
-    result->state = parameters[0]->state;
-}
-
-static std::unique_ptr<ScalarFunction> getStructExtractFunction(LogicalTypeID logicalTypeID) {
-    return std::make_unique<ScalarFunction>(StructExtractFunctions::name,
-        std::vector<LogicalTypeID>{logicalTypeID, LogicalTypeID::STRING}, LogicalTypeID::ANY,
-        nullptr, nullptr, StructExtractFunctions::compileFunc, StructExtractFunctions::bindFunc);
-}
-
-function_set StructExtractFunctions::getFunctionSet() {
-    function_set functions;
-    auto inputTypeIDs =
-        std::vector<LogicalTypeID>{LogicalTypeID::STRUCT, LogicalTypeID::NODE, LogicalTypeID::REL};
-    for (auto inputTypeID : inputTypeIDs) {
-        functions.push_back(getStructExtractFunction(inputTypeID));
-    }
     return functions;
 }
